Update of Danielas code to consider the new format and example files also in the core simulation + some grafical adaptations.
This commit is an update to Danielas code. The simulation data imports are use now the same example files. The new format is now also implemented in those files. Further, some grafical update were done to describe the new format. The example files for the forecast and the price was adapted to one day only and 15min interval.